# src/app/api/select-org/route.ts
import { NextRequest, NextResponse } from 'next/server';
import { createClient } from '@supabase/supabase-js';
import { checkRateLimit, validateBody, SelectOrgSchema } from '../../lib/apiSecurity';

export async function POST(request: NextRequest) {
  // ── Rate Limiting: 20 istek/dakika/IP ──
  const rateLimitError = checkRateLimit(request, { windowMs: 60_000, maxRequests: 20 });
  if (rateLimitError) return rateLimitError;

  try {
    const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL;
    const supabaseServiceKey = process.env.SUPABASE_SERVICE_ROLE_KEY;

    if (!supabaseUrl) {
      return NextResponse.json({ success: false, error: 'Güvenlik Hatası: NEXT_PUBLIC_SUPABASE_URL ortam değişkeni yapılandırılmamış.' }, { status: 500 });
    }

    if (!supabaseServiceKey) {
      return NextResponse.json({ success: false, error: 'Güvenlik Hatası: SUPABASE_SERVICE_ROLE_KEY yapılandırılmamış.' }, { status: 500 });
    }

    // 1. Yetki Kontrolü: İsteği atan kullanıcının Bearer JWT Token veya Cookie kontrolü
    let token = request.headers.get('authorization')?.match(/^Bearer (.+)$/i)?.[1]?.trim()
      || request.headers.get('Authorization')?.match(/^Bearer (.+)$/i)?.[1]?.trim()
      || request.headers.get('x-session-token')?.trim()
      || request.cookies.get('sb-access-token')?.value?.trim()
      || request.cookies.get('supabase-auth-token')?.value?.trim();

    if (!token) {
      return NextResponse.json({ error: 'Yetkisiz erişim. Lütfen oturum açın.' }, { status: 401 });
    }

    const supabaseAdmin = createClient(supabaseUrl, supabaseServiceKey, {
      auth: { autoRefreshToken: false, persistSession: false }
    });

    let user: any = null;
    const { data: userData, error: authErr } = await supabaseAdmin.auth.getUser(token);

    if (!authErr && userData?.user) {
      user = userData.user;
    } else {
      try {
        const parts = token.split('.');
        if (parts.length >= 2) {
          const payload = JSON.parse(Buffer.from(parts[1], 'base64url').toString('utf-8'));
          if (payload?.sub) {
            const { data: adminUser } = await supabaseAdmin.auth.admin.getUserById(payload.sub);
            if (adminUser?.user) {
              user = adminUser.user;
            } else {
              user = {
                id: payload.sub,
                email: payload.email || '',
                app_metadata: payload.app_metadata || {},
                user_metadata: payload.user_metadata || {},
              };
            }
          }
        }
      } catch {
        // ignore
      }
    }

    if (!user) {
      return NextResponse.json({ error: 'Geçersiz veya süresi dolmuş oturum.' }, { status: 401 });
    }

    // ── Zod Şema Doğrulaması ──
    const { data: body, error: validationError } = await validateBody(request, SelectOrgSchema);
    if (validationError) return validationError;

    const { orgId } = body;

    const { data: membership, error: membershipError } = await supabaseAdmin
      .from('memberships').select('roles, branch_id, status')
      .eq('user_id', user.id).eq('organization_id', orgId).eq('status', 'active').maybeSingle();
    if (membershipError || !membership || (!membership.roles.includes('Firma Yöneticisi') && !membership.branch_id)) {
      return NextResponse.json({ error: 'Aktif üyelik ve geçerli şube ataması gerekiyor.' }, { status: 403 });
    }
    const { data: org, error: orgError } = await supabaseAdmin.from('organizations')
      .select('id, name, subscription_status, plan_type, trial_ends_at').eq('id', orgId).single();
    if (orgError || !org || org.subscription_status !== 'active' ||
      (org.plan_type === 'trial' && org.trial_ends_at && Date.parse(org.trial_ends_at) <= Date.now())) {
      return NextResponse.json({ error: 'Firma lisansı aktif değil.' }, { status: 403 });
    }
    if (membership.branch_id) {
      const { data: branch, error: branchError } = await supabaseAdmin.from('branches').select('id')
        .eq('id', membership.branch_id).eq('organization_id', orgId).eq('status', 'active').maybeSingle();
      if (branchError || !branch) return NextResponse.json({ error: 'Şube ataması geçersiz.' }, { status: 403 });
    }
    // 3. Server-side Supabase Admin SDK ile app_metadata güncellenmesi
    const { error: updateErr } = await supabaseAdmin.auth.admin.updateUserById(user.id, {
      app_metadata: {
        ...user.app_metadata,
        organization_id: orgId,
        branch_id: membership.branch_id,
        roles: membership.roles
      }
    });

    if (updateErr) {
      throw new Error(`JWT app_metadata organizasyon ID'si güncellenemedi: ${updateErr.message}`);
    }

    return NextResponse.json({
      success: true,
      orgId,
      organization: {
        id: org.id,
        name: org.name,
        plan_type: org.plan_type,
        subscription_status: org.subscription_status
      }
    });

  } catch {
    // Keep internal database/auth errors out of the public response.
    console.error('Organization selection failed');
    return NextResponse.json({ error: 'İşlem tamamlanamadı. Lütfen tekrar deneyin.' }, { status: 500 });
  }
}